requirements.txt
cbor2
git+https://github.com/netd-tud/cbor4dns
dnspython
hvplot
jupyterlab
matplotlib
polars
pyarrow
sqlalchemy
